prysm/smart_contract/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
ape-foundry==0.8.9
ape-solidity==0.8.5
asttokens==2.4.1
attrs==25.3.0
base58==1.0.3
bitarray==3.6.1
black==25.1.0
cached-property==2.0.1
cchecksum==0.3.1
certifi==2025.8.3
charset-normalizer==3.4.3
circomlibpy==1.0.0
ckzg==2.1.1
click==8.1.8
cytoolz==1.0.1
dataclassy==0.11.1
decorator==5.2.1
eip712==0.2.13
eth-account==0.13.7
eth-ape==0.8.36
eth-bloom==3.1.0
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-pydantic-types==0.2.1
eth-rlp==2.2.0
eth-tester==0.13.0b1
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
ethpm-types==0.6.27
evm-trace==0.2.6
evmchains==0.1.5
executing==2.2.0
frozenlist==1.7.0
greenlet==3.2.4
hexbytes==1.3.1
idna==3.10
ijson==3.4.0
iniconfig==2.1.0
ipython==8.37.0
isort==6.0.1
jedi==0.19.2
lazyasd==0.1.4
lru-dict==1.3.0
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
morphys==1.0
msgspec==0.19.0
multidict==6.6.4
mypy==1.17.1
mypy_extensions==1.1.0
numpy==1.26.4
packaging==23.2
pandas==2.3.1
parsimonious==0.10.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-cid==0.3.0
py-ecc==8.0.0
py-evm==0.12.1b1
py-geth==5.6.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
py-solc-x==2.0.4
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
rich==13.9.4
rlp==4.1.0
safe-pysha3==1.0.5
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
stack-data==0.6.3
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
trie==3.1.0
types-requests==2.32.4.20250809
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
varint==1.0.2
watchdog==3.0.0
wcwidth==0.2.13
web3==7.13.0
websockets==15.0.1
yarl==1.20.1
